마이페이지 정보 변경 로직 수정 - member.getProfilePic()이 null일 경우 String.startsWith()에서 NPE 발생. null 여부도 같이 검증하도록 수정함.

+++ src/main/java/com/dongyang/dongpo/service/mypage/MyPageService.java
@@ -45,7 +45,7 @@ public class MyPageService {
 
         memberRepository.findByEmail(email).ifPresent(member -> {
             if (myPageUpdateDto.getProfilePic() != null && !myPageUpdateDto.getProfilePic().isBlank()) {
-                if (member.getProfilePic().startsWith(bucketFullUrl))
+                if (member.getProfilePic() != null && member.getProfilePic().startsWith(bucketFullUrl))
                     s3Service.deleteFile(member.getProfilePic()); // S3에 있는 기존 프로필 사진 삭제
                 member.setProfilePic(myPageUpdateDto.getProfilePic());
                 log.info("Updated Member profilePic : {}", member.getEmail());
